separate http_test.rs
