+++ test/extensions/dynamic_modules/test_data/rust/Cargo.toml
@@ -17,23 +17,28 @@ features = ["testing"]
 name = "no_op"
 path = "no_op.rs"
 crate-type = ["cdylib"]
+test = true
 
 [[example]]
 name = "no_program_init"
 path = "no_program_init.rs"
 crate-type = ["cdylib"]
+test = true
 
 [[example]]
 name = "program_init_fail"
 path = "program_init_fail.rs"
 crate-type = ["cdylib"]
+test = true
 
 [[example]]
 name = "abi_version_mismatch"
 path = "abi_version_mismatch.rs"
 crate-type = ["cdylib"]
+test = true
 
 [[example]]
 name = "http"
 path = "http.rs"
 crate-type = ["cdylib"]
+test = true
